requirements.txt
datasets
tqdm
pyyaml
torch
transformers
peft
accelerate
torchinfo
wandb
python-dotenv
matplotlib
imageio